+++ src/factory.ts
@@ -1,5 +1,4 @@
 /**
  * factory
  */
-// export * from '@pecorino/factory';
 export * from '@chevre/factory';